+++ libs/settings.py
@@ -1,10 +1,13 @@
 import pickle
 import os
+import sys
 
 class Settings(object):
     def __init__(self):
+        # Be default, the home will be in the same folder as labelImg
+        home = os.path.expanduser("~")
         self.data = {}
-        self.path = '.settings.pkl'
+        self.path = os.path.join(home, '.labelImgSettings.pkl')
 
     def __setitem__(self, key, value):
         self.data[key] = value
@@ -18,9 +21,10 @@ class Settings(object):
         return default
 
     def save(self):
-        with open(self.path, 'wb') as f:
-            pickle.dump(self.data, f, pickle.HIGHEST_PROTOCOL)
-            return True
+        if self.path:
+            with open(self.path, 'wb') as f:
+                pickle.dump(self.data, f, pickle.HIGHEST_PROTOCOL)
+                return True
         return False
 
     def load(self):
@@ -30,3 +34,9 @@ class Settings(object):
                 return True
         return False
 
+    def reset(self):
+        if os.path.exists(self.path):
+            os.remove(self.path)
+            print ('Remove setting pkl file ${0}'.format(self.path))
+        self.data = {}
+        self.path = None