+++ libs/shape.py
@@ -10,6 +10,7 @@ except ImportError:
     from PyQt4.QtCore import *
 
 from libs.lib import distance
+import sys
 
 DEFAULT_LINE_COLOR = QColor(0, 255, 0, 128)
 DEFAULT_FILL_COLOR = QColor(255, 0, 0, 128)
@@ -17,6 +18,7 @@ DEFAULT_SELECT_LINE_COLOR = QColor(255, 255, 255)
 DEFAULT_SELECT_FILL_COLOR = QColor(0, 128, 255, 155)
 DEFAULT_VERTEX_FILL_COLOR = QColor(0, 255, 0, 255)
 DEFAULT_HVERTEX_FILL_COLOR = QColor(255, 0, 0)
+MIN_Y_LABEL = 10
 
 
 class Shape(object):
@@ -36,12 +38,13 @@ class Shape(object):
     point_size = 8
     scale = 1.0
 
-    def __init__(self, label=None, line_color=None,difficult = False):
+    def __init__(self, label=None, line_color=None, difficult=False, paintLabel=False):
         self.label = label
         self.points = []
         self.fill = False
         self.selected = False
         self.difficult = difficult
+        self.paintLabel = paintLabel
 
         self._highlightIndex = None
         self._highlightMode = self.NEAR_VERTEX
@@ -67,9 +70,7 @@ class Shape(object):
         return False
 
     def addPoint(self, point):
-        if self.points and point == self.points[0]:
-            self.close()
-        else:
+        if not self.reachMaxPoints():
             self.points.append(point)
 
     def popPoint(self):
@@ -109,6 +110,25 @@ class Shape(object):
             painter.drawPath(line_path)
             painter.drawPath(vrtx_path)
             painter.fillPath(vrtx_path, self.vertex_fill_color)
+
+            # Draw text at the top-left
+            if self.paintLabel:
+                min_x = sys.maxsize
+                min_y = sys.maxsize
+                for point in self.points:
+                    min_x = min(min_x, point.x())
+                    min_y = min(min_y, point.y())
+                if min_x != sys.maxsize and min_y != sys.maxsize:
+                    font = QFont()
+                    font.setPointSize(8)
+                    font.setBold(True)
+                    painter.setFont(font)
+                    if(self.label == None):
+                        self.label = ""
+                    if(min_y < MIN_Y_LABEL):
+                        min_y += MIN_Y_LABEL
+                    painter.drawText(min_x, min_y, self.label)
+
             if self.fill:
                 color = self.select_fill_color if self.selected else self.fill_color
                 painter.fillPath(line_path, color)
@@ -172,7 +192,7 @@ class Shape(object):
             shape.line_color = self.line_color
         if self.fill_color != Shape.fill_color:
             shape.fill_color = self.fill_color
-        shape.difficult = self.difficult 
+        shape.difficult = self.difficult
         return shape
 
     def __len__(self):