+++ HISTORY.rst
@@ -1,14 +1,65 @@
 =======
 History
 =======
-1.4.0 (2017-0-7)
+
+1.8.0 (2018-10-21)
+------------------
+
+* Support drawing sqaure rect
+* Add item single click slot
+* Fix issues
+
+1.7.0 (2018-05-18)
+------------------
+
+* Support YOLO
+* Fix minor issues
+
+
+1.6.1 (2018-04-17)
+------------------
+
+* Fix issue
+
+1.6.0 (2018-01-29)
+------------------
+
+* Add more pre-defined labels
+* Show cursor pose in status bar
+* Fix minor issues
+
+1.5.2 (2017-10-24)
+------------------
+
+* Assig different colors to different lablels
+
+1.5.1 (2017-9-27)
+------------------
+
+* Show a autosaving dialog
+
+1.5.0 (2017-9-14)
+------------------
+
+* Fix the issues
+* Add feature: Draw a box easier
+
+
+1.4.3 (2017-08-09)
+------------------
+
+* Refactor setting
+* Fix the issues
+
+
+1.4.0 (2017-07-07)
 ------------------
 
 * Add feature: auto saving
 * Add feature: single class mode
 * Fix the issues
 
-1.3.4 (2017-07-7)
+1.3.4 (2017-07-07)
 ------------------
 
 * Fix issues and improve zoom-in